refactor: 데이터 호출을 각 컴포넌트에 하도록 구성 - 더미데이터 위치 수정
